ButtonBank Alpha; New OSC Panel
This commit introduces the SurgeButtonBank widget, which is
fully functional but has an incomplete draw look and feel.
Using that widget, we re-panel the SurgeOSC to be way more
compact and obvious. At the same time, fix the problem with
OSC type switches resulting in mis-rendered parameters (which
Closes #68).
